backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
faiss-cpu==1.8.0
openai==1.40.0
tiktoken==0.7.0
